UI: Colocate checkrun lens code
Change-type: patch
